Split file into frontend and backend
closes #8
